analitycs-service/requirements.txt
fastapi
uvicorn
psycopg2-binary
requests
pandas
python-jose

